+++ updates/1.3.0/update/update.sh
@@ -1,10 +1,24 @@
 #!/bin/sh
 
-#export V_MAJOR=3
-#export V_MINOR=14
-#export V_POINT=1
-#export CHECKSUM=376627f9f44142198a26123544c6505cf126b84199697fe436f6603de0b466a7
-#alpine-upgrade.sh
+if apkinfo mygpiod
+then
+    HAVE_LIBGPIOD=1
+    apk del libgpiod mygpiod
+else
+    HAVE_LIBGPIOD=0
+fi
+
+export V_MAJOR=3
+export V_MINOR=19
+export V_POINT=0
+export CHECKSUM=5621e7e597c3242605cd403a0a9109ec562892a6c8a185852b6b02ff88f5503c
+alpine-upgrade.sh
 
 # Save shutdowntime in apkovl
 lbu_include /var/lib/misc/openrc-shutdowntime
+
+if apkinfo libgpiod
+then
+    apk update
+    apk add mympdos-libgpiod2 mygpiod
+fi